move logic into EmojiTable
