+++ src/client/graphics/layers/EmojiTable.ts
@@ -1,5 +1,12 @@
 import { LitElement, css, html } from "lit";
 import { customElement, state } from "lit/decorators.js";
+import { EventBus } from "../../../core/EventBus";
+import { AllPlayers } from "../../../core/game/Game";
+import { GameView, PlayerView } from "../../../core/game/GameView";
+import { TerraNulliusImpl } from "../../../core/game/TerraNulliusImpl";
+import { ShowEmojiMenuEvent } from "../../InputHandler";
+import { SendEmojiIntentEvent } from "../../Transport";
+import { TransformHandler } from "../TransformHandler";
 
 const emojiTable: string[][] = [
   ["😀", "😊", "🥰", "😇", "😎"],
@@ -17,6 +24,10 @@ const emojiTable: string[][] = [
 
 @customElement("emoji-table")
 export class EmojiTable extends LitElement {
+  public eventBus: EventBus;
+  public transformHandler: TransformHandler;
+  public game: GameView;
+
   static styles = css`
     :host {
       display: block;
@@ -95,6 +106,35 @@ export class EmojiTable extends LitElement {
 
   @state()
   private _hidden = true;
+
+  initEventBus() {
+    this.eventBus.on(ShowEmojiMenuEvent, (e) => {
+      const cell = this.transformHandler.screenToWorldCoordinates(e.x, e.y);
+      if (!this.game.isValidCoord(cell.x, cell.y)) {
+        return;
+      }
+
+      const tile = this.game.ref(cell.x, cell.y);
+      if (!this.game.hasOwner(tile)) {
+        return;
+      }
+
+      const targetPlayer = this.game.owner(tile);
+      // maybe redundant due to owner check but better safe than sorry
+      if (targetPlayer instanceof TerraNulliusImpl) {
+        return;
+      }
+
+      this.showTable((emoji) => {
+        const recipient =
+          targetPlayer == this.game.myPlayer()
+            ? AllPlayers
+            : (targetPlayer as PlayerView);
+        this.eventBus.emit(new SendEmojiIntentEvent(recipient, emoji));
+        this.hideTable();
+      });
+    });
+  }
 
   private onEmojiClicked: (emoji: string) => void = () => {};
